What to do with bad column names?
How should students experience bad data files?

A bad column name cannot be mapped to SQL.

One solution is to store the string, spaces and all, and then create a hashmap to the actual table name.

This indirection might be complex and/or fragile.

Another is to rename the columns, and alert. That is what I've done here.

I need to think about this more.