Fixed a bug of sorts.
The histogram module needs a lot of work.

Also, the use of threading is problematic. I think it rewrites some things that seriously slow down this library.

But, I'll leave the example for now. It does not work well with my "compute" function. It's a factor of a bazillion slower with threading.